\name{readBam}
\alias{readBam}
\title{
  Read BAM (Binary Alignment/Map File)
}
\description{
  Reads Ion Torrent BAM files, which contain read alignment results.
}
\usage{
  readBam(
    bamFile,
  )
}
\arguments{
  \item{bamFile}{
    Name of the bam file to load
  }
}
\value{
}
\author{
  Eugene Ingerman
}
