\name{torrentR-package}
\alias{torrentR-package}
\docType{package}
\title{
  A package for handling data from Ion Torrent Systems
}
\description{
  This package is for exploratory analysis and visualization of data from Ion Torrent Systems.  Functionality is provided
  for loading processed signal data (from "1.wells" files) and for perfoming various subsequent anlayses and visualizations
  based.  There are plans to expand the package in the future to also handle data from the raw image files from which the
  signals are derived.
}
\details{
  \tabular{ll}{
    Package: \tab torrentR\cr
    Type: \tab Package\cr
    Version: \tab 0.4.1\cr
    Date: \tab 2011-06-20\cr
    License: \tab GPL-2\cr
    LazyLoad: \tab yes\cr
  }
}
\author{
  Simon Cawley

  Maintainer: <scawley@iontorrent.com>
}
\keyword{ package }
