\name{torrentR_version}
\alias{torrentR_version}
\title{
  Returns the version of the torrentR package
}
\description{
  Returns the version of the torrentR package
}
\usage{
  torrentR_version()
}
\value{
  Returns a character string specifying the torrentR version
}
\author{
  Simon Cawley
}
\seealso{
  \code{\link{plot_version}},
}
\examples{
  x <- torrentR_version()
}
